+++ aiokafka/protocol/types.py
@@ -19,6 +19,10 @@ T = TypeVar("T")
 
 ValueT: TypeAlias = Union[type[AbstractType[Any]], "String", "Array", "Schema"]
 
+TaggedFieldId: TypeAlias = tuple[str, int]
+
+FieldId: TypeAlias = TaggedFieldId | str
+
 
 def _pack(f: Callable[[T], bytes], value: T) -> bytes:
     try:
@@ -47,11 +51,11 @@ class Int8(AbstractType[int]):
     _unpack = struct.Struct(">b").unpack
 
     @classmethod
-    def encode(cls, value: int) -> bytes:
+    def encode(cls, value: int, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> int:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> int:
         return _unpack(cls._unpack, data.read(1))
 
 
@@ -60,11 +64,11 @@ class Int16(AbstractType[int]):
     _unpack = struct.Struct(">h").unpack
 
     @classmethod
-    def encode(cls, value: int) -> bytes:
+    def encode(cls, value: int, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> int:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> int:
         return _unpack(cls._unpack, data.read(2))
 
 
@@ -73,11 +77,11 @@ class Int32(AbstractType[int]):
     _unpack = struct.Struct(">i").unpack
 
     @classmethod
-    def encode(cls, value: int) -> bytes:
+    def encode(cls, value: int, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> int:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> int:
         return _unpack(cls._unpack, data.read(4))
 
 
@@ -86,11 +90,11 @@ class UInt32(AbstractType[int]):
     _unpack = struct.Struct(">I").unpack
 
     @classmethod
-    def encode(cls, value: int) -> bytes:
+    def encode(cls, value: int, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> int:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> int:
         return _unpack(cls._unpack, data.read(4))
 
 
@@ -99,11 +103,11 @@ class Int64(AbstractType[int]):
     _unpack = struct.Struct(">q").unpack
 
     @classmethod
-    def encode(cls, value: int) -> bytes:
+    def encode(cls, value: int, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> int:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> int:
         return _unpack(cls._unpack, data.read(8))
 
 
@@ -112,26 +116,39 @@ class Float64(AbstractType[float]):
     _unpack = struct.Struct(">d").unpack
 
     @classmethod
-    def encode(cls, value: float) -> bytes:
+    def encode(cls, value: float, flexible: bool = False) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> float:
+    def decode(cls, data: BytesIO, flexible: bool = False) -> float:
         return _unpack(cls._unpack, data.read(8))
 
 
 class String:
-    def __init__(self, encoding: str = "utf-8"):
+    def __init__(self, encoding: str = "utf-8", allow_flexible: bool = True):
         self.encoding = encoding
+        self.allow_flexible = allow_flexible
 
-    def encode(self, value: str | None) -> bytes:
+    def encode(self, value: str | None, flexible: bool) -> bytes:
         if value is None:
-            return Int16.encode(-1)
+            return (
+                UnsignedVarInt32.encode(0)
+                if flexible and self.allow_flexible
+                else Int16.encode(-1, flexible)
+            )
         encoded_value = str(value).encode(self.encoding)
-        return Int16.encode(len(encoded_value)) + encoded_value
+        return (
+            UnsignedVarInt32.encode(len(encoded_value) + 1) + encoded_value
+            if flexible and self.allow_flexible
+            else Int16.encode(len(encoded_value), flexible) + encoded_value
+        )
 
-    def decode(self, data: BytesIO) -> str | None:
-        length = Int16.decode(data)
+    def decode(self, data: BytesIO, flexible: bool) -> str | None:
+        length = (
+            UnsignedVarInt32.decode(data) - 1
+            if flexible and self.allow_flexible
+            else Int16.decode(data, flexible)
+        )
         if length < 0:
             return None
         value = data.read(length)
@@ -146,15 +163,25 @@ class String:
 
 class Bytes(AbstractType[bytes | None]):
     @classmethod
-    def encode(cls, value: bytes | None) -> bytes:
+    def encode(cls, value: bytes | None, flexible: bool) -> bytes:
         if value is None:
-            return Int32.encode(-1)
+            return (
+                UnsignedVarInt32.encode(0) if flexible else Int32.encode(-1, flexible)
+            )
         else:
-            return Int32.encode(len(value)) + value
+            return (
+                UnsignedVarInt32.encode(len(value) + 1) + value
+                if flexible
+                else Int32.encode(len(value), flexible) + value
+            )
 
     @classmethod
-    def decode(cls, data: BytesIO) -> bytes | None:
-        length = Int32.decode(data)
+    def decode(cls, data: BytesIO, flexible: bool) -> bytes | None:
+        length = (
+            UnsignedVarInt32.decode(data) - 1
+            if flexible
+            else Int32.decode(data, flexible)
+        )
         if length < 0:
             return None
         value = data.read(length)
@@ -174,33 +201,94 @@ class Boolean(AbstractType[bool]):
     _unpack = struct.Struct(">?").unpack
 
     @classmethod
-    def encode(cls, value: bool) -> bytes:
+    def encode(cls, value: bool, flexible: bool) -> bytes:
         return _pack(cls._pack, value)
 
     @classmethod
-    def decode(cls, data: BytesIO) -> bool:
+    def decode(cls, data: BytesIO, flexible: bool) -> bool:
         return _unpack(cls._unpack, data.read(1))
 
 
 class Schema:
     names: tuple[str, ...]
+    tags: tuple[int, ...]
     fields: tuple[ValueT, ...]
 
-    def __init__(self, *fields: tuple[str, ValueT]):
+    def __init__(self, *fields: tuple[FieldId, ValueT]):
         if fields:
-            self.names, self.fields = zip(*fields, strict=False)
+            tagged_names, values = zip(
+                *(
+                    (key, value) if isinstance(key, tuple) else ((key, None), value)
+                    for key, value in fields
+                ),
+                strict=False,
+            )
+            self.names = tuple(name for name, _ in tagged_names)
+            self.tags = tuple(tag for _, tag in tagged_names)
+            self.fields = tuple(values)
         else:
-            self.names, self.fields = (), ()
+            self.names, self.tags, self.fields = (), (), ()
 
-    def encode(self, item: Sequence[Any]) -> bytes:
+    def encode(self, item: Sequence[Any], flexible: bool) -> bytes:
         if len(item) != len(self.fields):
             raise ValueError("Item field count does not match Schema")
-        return b"".join(field.encode(item[i]) for i, field in enumerate(self.fields))
+        return b"".join(
+            field.encode(item[i], flexible)
+            for i, field in enumerate(self.fields)
+            if self.tags[i] is None
+        ) + (
+            self._encode_tagged_fields(
+                {
+                    self.tags[i]: field.encode(item[i], flexible)
+                    for i, field in enumerate(self.fields)
+                    if self.tags[i] is not None
+                }
+            )
+            if flexible
+            else b""
+        )
 
     def decode(
-        self, data: BytesIO
+        self, data: BytesIO, flexible: bool
     ) -> tuple[Any | str | None | list[Any | tuple[Any, ...]], ...]:
-        return tuple(field.decode(data) for field in self.fields)
+        result = [
+            field.decode(data, flexible) if self.tags[i] is None else None
+            for i, field in enumerate(self.fields)
+        ]
+        if flexible:
+            tagged_fields = self._decode_tagged_fields(data)
+            for i, tag in enumerate(self.tags):
+                if tag is not None:
+                    encoded_value = tagged_fields.get(tag)
+                    if encoded_value is not None:
+                        result[i] = self.fields[i].decode(
+                            BytesIO(encoded_value), flexible
+                        )
+
+        return tuple(result)
+
+    @staticmethod
+    def _encode_tagged_fields(value: dict[int, bytes]) -> bytes:
+        ret = UnsignedVarInt32.encode(len(value))
+        for k, v in value.items():
+            assert isinstance(k, int) and k >= 0, f"Key {k} is not a positive integer"
+            ret += UnsignedVarInt32.encode(k)
+            ret += UnsignedVarInt32.encode(len(v))
+            ret += v
+        return ret
+
+    @staticmethod
+    def _decode_tagged_fields(data: BytesIO) -> dict[int, bytes]:
+        num_fields = UnsignedVarInt32.decode(data)
+        ret: dict[int, bytes] = {}
+        if not num_fields:
+            return ret
+        for _ in range(num_fields):
+            tag = UnsignedVarInt32.decode(data)
+            size = UnsignedVarInt32.decode(data)
+            val = data.read(size)
+            ret[tag] = val
+        return ret
 
     def __len__(self) -> int:
         return len(self.fields)
@@ -227,16 +315,18 @@ class Array:
 
     @overload
     def __init__(
-        self, array_of_0: tuple[str, ValueT], *array_of: tuple[str, ValueT]
+        self,
+        array_of_0: tuple[FieldId, ValueT],
+        *array_of: tuple[FieldId, ValueT],
     ): ...
 
     def __init__(
         self,
-        array_of_0: ValueT | tuple[str, ValueT],
-        *array_of: tuple[str, ValueT],
+        array_of_0: ValueT | tuple[FieldId, ValueT],
+        *array_of: tuple[FieldId, ValueT],
     ) -> None:
         if array_of:
-            array_of_0 = cast(tuple[str, ValueT], array_of_0)
+            array_of_0 = cast(tuple[FieldId, ValueT], array_of_0)
             self.array_of = Schema(array_of_0, *array_of)
         else:
             array_of_0 = cast(ValueT, array_of_0)
@@ -247,19 +337,33 @@ class Array:
             else:
                 raise ValueError("Array instantiated with no array_of type")
 
-    def encode(self, items: Sequence[Any] | None) -> bytes:
+    def encode(self, items: Sequence[Any] | None, flexible: bool) -> bytes:
         if items is None:
-            return Int32.encode(-1)
-        encoded_items = (self.array_of.encode(item) for item in items)
-        return b"".join(
-            (Int32.encode(len(items)), *encoded_items),
+            return (
+                UnsignedVarInt32.encode(0) if flexible else Int32.encode(-1, flexible)
+            )
+        encoded_items = (self.array_of.encode(item, flexible) for item in items)
+        return (
+            b"".join(
+                (UnsignedVarInt32.encode(len(items) + 1), *encoded_items),
+            )
+            if flexible
+            else b"".join(
+                (Int32.encode(len(items), flexible), *encoded_items),
+            )
         )
 
-    def decode(self, data: BytesIO) -> list[Any | tuple[Any, ...]] | None:
-        length = Int32.decode(data)
+    def decode(
+        self, data: BytesIO, flexible: bool
+    ) -> list[Any | tuple[Any, ...]] | None:
+        length = (
+            UnsignedVarInt32.decode(data) - 1
+            if flexible
+            else Int32.decode(data, flexible)
+        )
         if length == -1:
             return None
-        return [self.array_of.decode(data) for _ in range(length)]
+        return [self.array_of.decode(data, flexible) for _ in range(length)]
 
     def repr(self, list_of_items: Sequence[Any] | None) -> str:
         if list_of_items is None:
@@ -267,7 +371,7 @@ class Array:
         return "[" + ", ".join(self.array_of.repr(item) for item in list_of_items) + "]"
 
 
-class UnsignedVarInt32(AbstractType[int]):
+class UnsignedVarInt32:
     @classmethod
     def decode(cls, data: BytesIO) -> int:
         value, i = 0, 0
@@ -293,128 +397,3 @@ class UnsignedVarInt32(AbstractType[int]):
             value >>= 7
         ret += struct.pack("B", value)
         return ret
-
-
-class VarInt32(AbstractType[int]):
-    @classmethod
-    def decode(cls, data: BytesIO) -> int:
-        value = UnsignedVarInt32.decode(data)
-        return (value >> 1) ^ -(value & 1)
-
-    @classmethod
-    def encode(cls, value: int) -> bytes:
-        # bring it in line with the java binary repr
-        value &= 0xFFFFFFFF
-        return UnsignedVarInt32.encode((value << 1) ^ (value >> 31))
-
-
-class VarInt64(AbstractType[int]):
-    @classmethod
-    def decode(cls, data: BytesIO) -> int:
-        value, i = 0, 0
-        b: int
-        while True:
-            (b,) = struct.unpack("B", data.read(1))
-            if not (b & 0x80):
-                break
-            value |= (b & 0x7F) << i
-            i += 7
-            if i > 63:
-                raise ValueError(f"Invalid value {value}")
-        value |= b << i
-        return (value >> 1) ^ -(value & 1)
-
-    @classmethod
-    def encode(cls, value: int) -> bytes:
-        # bring it in line with the java binary repr
-        value &= 0xFFFFFFFFFFFFFFFF
-        v = (value << 1) ^ (value >> 63)
-        ret = b""
-        while (v & 0xFFFFFFFFFFFFFF80) != 0:
-            b = (value & 0x7F) | 0x80
-            ret += struct.pack("B", b)
-            v >>= 7
-        ret += struct.pack("B", v)
-        return ret
-
-
-class CompactString(String):
-    def decode(self, data: BytesIO) -> str | None:
-        length = UnsignedVarInt32.decode(data) - 1
-        if length < 0:
-            return None
-        value = data.read(length)
-        if len(value) != length:
-            raise ValueError("Buffer underrun decoding string")
-        return value.decode(self.encoding)
-
-    def encode(self, value: str | None) -> bytes:
-        if value is None:
-            return UnsignedVarInt32.encode(0)
-        encoded_value = str(value).encode(self.encoding)
-        return UnsignedVarInt32.encode(len(encoded_value) + 1) + encoded_value
-
-
-class TaggedFields(AbstractType[dict[int, bytes]]):
-    @classmethod
-    def decode(cls, data: BytesIO) -> dict[int, bytes]:
-        num_fields = UnsignedVarInt32.decode(data)
-        ret: dict[int, bytes] = {}
-        if not num_fields:
-            return ret
-        prev_tag = -1
-        for _ in range(num_fields):
-            tag = UnsignedVarInt32.decode(data)
-            if tag <= prev_tag:
-                raise ValueError(f"Invalid or out-of-order tag {tag}")
-            prev_tag = tag
-            size = UnsignedVarInt32.decode(data)
-            val = data.read(size)
-            ret[tag] = val
-        return ret
-
-    @classmethod
-    def encode(cls, value: dict[int, bytes]) -> bytes:
-        ret = UnsignedVarInt32.encode(len(value))
-        for k, v in value.items():
-            # do we allow for other data types ?? It could get complicated really fast
-            assert isinstance(v, bytes), f"Value {v!r} is not a byte array"
-            assert isinstance(k, int) and k > 0, f"Key {k} is not a positive integer"
-            ret += UnsignedVarInt32.encode(k)
-            ret += v
-        return ret
-
-
-class CompactBytes(AbstractType[bytes | None]):
-    @classmethod
-    def decode(cls, data: BytesIO) -> bytes | None:
-        length = UnsignedVarInt32.decode(data) - 1
-        if length < 0:
-            return None
-        value = data.read(length)
-        if len(value) != length:
-            raise ValueError("Buffer underrun decoding Bytes")
-        return value
-
-    @classmethod
-    def encode(cls, value: bytes | None) -> bytes:
-        if value is None:
-            return UnsignedVarInt32.encode(0)
-        else:
-            return UnsignedVarInt32.encode(len(value) + 1) + value
-
-
-class CompactArray(Array):
-    def encode(self, items: Sequence[Any] | None) -> bytes:
-        if items is None:
-            return UnsignedVarInt32.encode(0)
-        encoded_items = (self.array_of.encode(item) for item in items)
-        return b"".join(
-            (UnsignedVarInt32.encode(len(items) + 1), *encoded_items),
-        )
-
-    def decode(self, data: BytesIO) -> list[Any | tuple[Any, ...]] | None:
-        length = UnsignedVarInt32.decode(data) - 1
-        if length == -1:
-            return None
-        return [self.array_of.decode(data) for _ in range(length)]